modified: move request to models folder
