+++ SP25_NET1721_RPR231_PRJ_G1_SkincareProductSalesSystem/SkincareProductSalesSystem.Services/Models/SkinTestModels/UpdateSkinTestRequest.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SkincareProductSalesSystem.Services.Models.SkinTestModels
+{
+    public class UpdateSkinTestRequest
+    {
+        public string Question { get; set; } = string.Empty;
+
+        public bool IsActive { get; set; } = false;
+
+        public int? QuestionOrder { get; set; }
+    }
+}
