+++ SP25_NET1721_RPR231_PRJ_G1_SkincareProductSalesSystem/SkincareProductSalesSystem.Services/SkinTestService.cs
@@ -1,6 +1,7 @@
 ﻿using SkincareProductSalesSystem.Repositories;
 using SkincareProductSalesSystem.Repositories.Models;
 using SkincareProductSalesSystem.Services.Base;
+using SkincareProductSalesSystem.Services.Models.SkinTestModels;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
@@ -133,27 +134,6 @@ namespace SkincareProductSalesSystem.Services
 
 }
 
-public class UpdateSkinTestRequest
-{
-    public string Question { get; set; } = string.Empty;
 
-    public bool IsActive { get; set; } = false;
 
-    public int? QuestionOrder { get; set; }
-}
-
-public class CreateSkinTestRequest
-{
-    public string Question { get; set; } = string.Empty;
-
-    public int? QuestionOrder { get; set; }
-
-    public List<CreateSkinTestOptionRequest> Options { get; set; }
-}
-public class CreateSkinTestOptionRequest
-{
-    public string OptionText { get; set; }
-
-    public string SkinTypeId { get; set; }
-}
